clean up folder structure and verify route
