+++ src/controllers/apiControllers/index.ts
@@ -1,0 +1,6 @@
+export * from "./itemsController";
+export * from "./AuthController";
+export * from "./userController";
+export * from "./tagController";
+export * from "./SlugController";
+export * from "./notFoundController";
