+++ src/controllers/apiControllers/notFoundController.ts
@@ -1,4 +1,4 @@
-import { NotFoundError } from "../errors";
+import { NotFoundError } from "../../errors";
 
 export const notFound = () => {
   throw new NotFoundError("The requested endpoint doesn't exist");
