+++ src/controllers/viewControllers/index.ts
@@ -1,0 +1,4 @@
+export * from "./itemViewControllers";
+export * from "./authViewControllers";
+export * from "./userViewControllers";
+export * from "./homeViewControllers";
