+++ src/views/forbidden.handlebars
@@ -25,7 +25,7 @@
   
   <div class="action-buttons">
     <a href="/" class="back-home">Back to Home</a>
-    <a href="/login" class="login-btn">Log In</a>
+    <a href="/auth/login" class="login-btn">Log In</a>
   </div>
 
   {{#if message}}
