+++ src/views/home.handlebars
@@ -72,7 +72,7 @@
     <div class="grid-section">
       <h2 class="section-title">Account</h2>
       <div class="tile-group">
-        <a href="/logout" class="tile">
+        <a href="/auth/logout" class="tile">
           <div class="tile-icon">
             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
               <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path>
