+++ src/views/layouts/main.handlebars
@@ -21,7 +21,7 @@
             </div>
           {{/if}}
           {{#if user}}
-            <a href="/logout">Logout</a>
+            <a href="/auth/logout">Logout</a>
           {{/if}}
         </div>
       </nav>
