+++ src/views/login.handlebars
@@ -27,6 +27,6 @@
     <button type="submit" class="login-button">Login</button>
   </form>
   <div class="register-link">
-    <p>Don't have an account? <a href="/register">Register here</a></p>
+    <p>Don't have an account? <a href="/auth/register">Register here</a></p>
   </div>
 </div>
