+++ src/views/register.handlebars
@@ -35,7 +35,7 @@
     </div>
     <h2>Registration Successful!</h2>
     <div id="success-message" class="success-message"></div>
-    <a href="/login" class="login-button">Go to Login</a>
+    <a href="/auth/login" class="login-button">Go to Login</a>
   </div>
   
   <form id="register-form" novalidate>
@@ -117,6 +117,6 @@
   </form>
   
   <div class="login-link">
-    <p>Already have an account? <a href="/login">Login here</a></p>
+    <p>Already have an account? <a href="/auth/login">Login here</a></p>
   </div>
 </div>
